move tags commands and test to a tags folder
