+++ gitblend/cli.py
@@ -1,14 +1,9 @@
 import argparse
 import sys
 
-from gitblend.commands import (
-    create_tag,
-    delete_tag,
-    list_tags,
-    rename_tag,
-    update_all,
-    version,
-)
+from gitblend.commands import update_all, version
+from gitblend.commands.tags import (create_tag, delete_tag, list_tags,
+                                    rename_tag)
 
 
 def add_create_tag_command(subparsers):
